gherkin-image/requirements.txt
Flask
easyocr
google-generativeai
python-dotenv
werkzeug
